ecourts-scraper/requirements.txt
fastapi
uvicorn[standard]
redis
python-dotenv
requests
beautifulsoup4
bleach
bleach[css]
pillow
pytesseract
jinja2
python-multipart
httpx
fakeredis
pydantic-settings
tinycss2